added tax formula for highincome salary and changed salary's data type to double to get result more appropriate

+++ app/src/main/java/com/example/taxcalculator/MainActivity.java
@@ -32,8 +32,8 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
 
             TaxCalculator taxc=new TaxCalculator();
             taxc.setAmount(salary);
-            float tax=taxc.taxableAmount();
-            tvTaxableAmount.setText(Float.toString(tax));
+            Double tax=taxc.taxableAmount();
+            tvTaxableAmount.setText(Double.toString(tax));
         }
